Move CsvSchema to tests
